feat: add encode api url
