+++ app/services/loklok/utils.server.ts
@@ -1,6 +1,7 @@
+import encode from '~/utils/encode';
 import { lruCache } from '../lru-cache';
 
-export const LOKLOK_URL = 'https://loklok.vercel.app/api';
+export const LOKLOK_URL = encode(process.env.LOKLOK_API_URL || '');
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const fetcher = async <T = any>(url: string): Promise<T> => {
